renamed the files & organized the codes
