+++ graze/__init__.py
@@ -1,4 +1,3 @@
-from .youtube.base import Youtube as youtube
-from .britannica.main import Britannica as britannica
-from .wikipedia.main import WikiScraper as wikipedia
-from .unsplash import unsplash
+from .queries import Queries
+from .utils import *
+from ._transcripts import Youtube
