--- graze/britannica/search_queries.json
@@ -1,41 +1,0 @@
-[
-  "antarctica",
-  "colonization",
-  "world war",
-  "asia",
-  "africa",
-  "australia",
-  "holocaust",
-  "voyages",
-  "biological viruses",
-  "Martin Luther King Jr",
-  "Abraham Lincon",
-  "Quarks",
-  "Quantum Mechanincs",
-  "Biological Viruses",
-  "Drugs",
-  "Rockets",
-  "Physics",
-  "Mathematics",
-  "nuclear physics",
-  "nuclear fusion",
-  "CRISPR CAS-9",
-  "virginia woolf",
-  "cocaine", 
-  "marijuana",
-  "apollo missions",
-  "birds",
-  "blogs",
-  "journal",
-  "Adolf Hitler",
-  "Presidents of United States",
-  "genders and sexes",
-  "journalism",
-  "maths theories",
-  "matter and particles",
-  "discoveries",
-  "authoers and writers",
-  "poets and novel writers",
-  "literature",
-  "awards and honors"
-]
