--- graze/imageQuery.json
@@ -1,104 +1,0 @@
-[
-  "american football",
-  "india",
-  "indian festivals",
-  "cars",
-  "street",
-  "street photography",
-  "wild animals",
-  "politics",
-  "mountains",
-  "rivers",
-  "sunsets",
-  "sunrises",
-  "forests",
-  "ocean",
-  "beaches",
-  "architecture",
-  "bridges",
-  "castles",
-  "cityscapes",
-  "clouds",
-  "deserts",
-  "flowers",
-  "gardens",
-  "landscapes",
-  "night sky",
-  "waterfalls",
-  "birds",
-  "fish",
-  "reptiles",
-  "insects",
-  "trees",
-  "autumn",
-  "winter",
-  "spring",
-  "summer",
-  "abstract art",
-  "black and white",
-  "portraits",
-  "macro photography",
-  "minimalism",
-  "patterns",
-  "space",
-  "planets",
-  "stars",
-  "galaxies",
-  "technology",
-  "robots",
-  "drones",
-  "sports",
-  "yoga",
-  "fitness",
-  "cycling",
-  "running",
-  "hiking",
-  "camping",
-  "travel",
-  "historical sites",
-  "monuments",
-  "museums",
-  "food",
-  "drinks",
-  "desserts",
-  "coffee",
-  "tea",
-  "wine",
-  "cocktails",
-  "fashion",
-  "clothing",
-  "shoes",
-  "accessories",
-  "makeup",
-  "hairstyles",
-  "tattoos",
-  "piercings",
-  "jewelry",
-  "watches",
-  "books",
-  "libraries",
-  "education",
-  "schools",
-  "universities",
-  "graduation",
-  "science",
-  "laboratories",
-  "microscopes",
-  "experiments",
-  "meditation",
-  "relaxation",
-  "spa",
-  "wellness",
-  "interior design",
-  "furniture",
-  "home decor",
-  "gardening",
-  "pets",
-  "dogs",
-  "cats",
-  "horses",
-  "farm animals",
-  "zoo",
-  "circus",
-  "theater"
-]
