+++ graze/topics.json
@@ -151,5 +151,32 @@
     "UCKWaEZ-_VweaEx1j62do_vQ", "UCFQDtftsHGzSh1-TReNT4lA", "UC6biysICWOJ-C3P4Tyeggzg",
     "UCFe-pfe0a9bDvWy74Jd7vFg", "UCUgZq9PkDp1xaEivtcfJPSg", "UC0-7PyfpOIJpNyi8WrHiyXA",
     "UCPb7xe-MQ0KiJpaKBWFZtTA", "UCycGV6fAhD_-7GPmCkkESdw", "UCG7J20LhUeLl6y_Emi7OJrA"
+  ],
+  "image_topics": [ "american football", "india", "europe",
+    "indian festivals", "cars", "street", "street photography",
+    "wild animals", "politics", "mountains", "rivers",
+    "sunsets", "sunrises", "forests", "ocean",
+    "beaches", "architecture", "bridges", "castles",
+    "cityscapes", "clouds", "deserts", "flowers",
+    "gardens", "landscapes", "night sky", "waterfalls",
+    "birds", "fish", "reptiles", "insects",
+    "trees", "autumn", "winter", "spring",
+    "summer", "abstract art", "black and white", "portraits",
+    "macro photography", "minimalism", "patterns", "space",
+    "planets", "stars", "galaxies", "technology",
+    "robots", "drones", "sports", "yoga",
+    "fitness", "cycling", "running", "hiking",
+    "camping", "travel", "historical sites", "monuments",
+    "museums", "food", "drinks", "desserts",
+    "coffee", "tea", "wine", "cocktails",
+    "fashion", "clothing", "shoes", "accessories",
+    "makeup", "hairstyles", "tattoos", "piercings",
+    "jewelry", "watches", "books", "libraries",
+    "education", "schools", "universities", "graduation",
+    "science", "laboratories", "microscopes", "experiments",
+    "meditation", "relaxation", "spa", "wellness",
+    "interior design", "furniture", "home decor", "gardening",
+    "pets", "dogs", "cats", "horses",
+    "farm animals", "zoo", "circus", "theater"
   ]
 }
