--- webgraze/base.py
@@ -1,1 +1,0 @@
-from .youtube._transcripts import Youtube
